+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### Pre - [0.0.2]
+
+- refactor: checkbox component
+  
 ### Pre - [0.0.1]
 
 - initial package setup